Move tree.makeExampleScore
create tree.examples for keeping makeExampleScore

Just some cleanup.  Not sure tree will survive m21 v10
