+++ music21/tree/core.py
@@ -65,7 +65,7 @@ class AVLNode(common.SlottedObjectMixin):
 
         This attribute is used to help balance the AVL tree.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...                    classList=(note.Note, chord.Chord))
         >>> print(scoreTree.debug())
@@ -111,7 +111,7 @@ class AVLNode(common.SlottedObjectMixin):
 
         This property is used to help balance the AVL tree.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...              classList=(note.Note, chord.Chord))
         >>> print(scoreTree.debug())
@@ -149,7 +149,7 @@ class AVLNode(common.SlottedObjectMixin):
         The position of this node -- this is often the same as the offset of
         the node in a containing score, but does not need to be. It could be the .sortTuple
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...            classList=(note.Note, chord.Chord))
         >>> print(scoreTree.rootNode.debug())
@@ -176,7 +176,7 @@ class AVLNode(common.SlottedObjectMixin):
 
         After setting the left child you need to do a node update. with node.update()
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...           classList=(note.Note, chord.Chord))
         >>> print(scoreTree.rootNode.debug())
@@ -199,7 +199,7 @@ class AVLNode(common.SlottedObjectMixin):
 
         After setting the right child you need to do a node update. with node.update()
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...             classList=(note.Note, chord.Chord))
         >>> print(scoreTree.rootNode.debug())
@@ -275,7 +275,7 @@ class AVLNode(common.SlottedObjectMixin):
         '''
         Get a debug of the Node:
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...              classList=(note.Note, chord.Chord))
         >>> rn = scoreTree.rootNode
@@ -297,7 +297,7 @@ class AVLNode(common.SlottedObjectMixin):
 
         Called recursively
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...            classList=(note.Note, chord.Chord))
         >>> rn = scoreTree.rootNode
@@ -338,7 +338,7 @@ class AVLNode(common.SlottedObjectMixin):
 
         We create a score with everything correct.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...             classList=(note.Note, chord.Chord))
         >>> n = scoreTree.rootNode
