+++ music21/tree/trees.py
@@ -123,7 +123,7 @@ class ElementTree(core.AVLTree):
         If element.sortTuple(self.source) returns the right information, it's a fast
         O(log n) search. If not his is an O(n log n) operation in python not C, so slow.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> lastNote = score.flatten().notes[-1]
         >>> lastNote in scoreTree
@@ -178,7 +178,7 @@ class ElementTree(core.AVLTree):
         (O(log n)), but it's O(log n) in Python while normal list slicing is O(1) in C, so
         don't use trees for __getitem__ searching if you don't have to.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> scoreTree
         <ElementTree {20} (0.0 <0.-25...> to 8.0) <music21.stream.Score exampleScore>>
@@ -243,7 +243,7 @@ class ElementTree(core.AVLTree):
         Gets the length of the ElementTree, i.e., the number of elements enclosed.
         This is a very very fast O(1).
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> len(scoreTree)
         20
@@ -300,7 +300,7 @@ class ElementTree(core.AVLTree):
         Sets elements at index `i` to `new`, but keeping the old position
         of the element there. (This is different from OffsetTrees, where things can move around).
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> n = scoreTree[10]
         >>> n
@@ -349,7 +349,7 @@ class ElementTree(core.AVLTree):
 
         Slow: O(n log n) time, but it's just for debugging
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> print(scoreTree)
         <ElementTree {20} (0.0 <0.-25...> to 8.0) <music21.stream.Score exampleScore>>
@@ -390,7 +390,7 @@ class ElementTree(core.AVLTree):
         Not an especially efficient way of using this beautiful tree object!  But useful
         for debugging or a final iteration for conversion.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> for x in scoreTree:
         ...     print(x)
@@ -582,7 +582,7 @@ class ElementTree(core.AVLTree):
 
         See __getitem__ for caveats about speed.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> scoreTree
         <ElementTree {20} (0.0 <0.-25...> to 8.0) <music21.stream.Score exampleScore>>
@@ -658,7 +658,7 @@ class ElementTree(core.AVLTree):
 
         Slow: O(n log n) time so don't make this your main thing.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True)
         >>> scoreTree
         <ElementTree {20} (0.0 <0.-25...> to 8.0) <music21.stream.Score exampleScore>>
@@ -699,7 +699,7 @@ class ElementTree(core.AVLTree):
 
         If the element is in the original score, then it should be very fast (O(log n))
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreFlat = score.flatten()
         >>> n = scoreFlat.notes[-1]
 
@@ -823,7 +823,7 @@ class ElementTree(core.AVLTree):
         r'''
         Gets the earliest position in this tree.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> elTree = score.asTree()
         >>> elTree.lowestPosition().shortRepr()
         '0.0 <0.-20...>'
@@ -848,7 +848,7 @@ class ElementTree(core.AVLTree):
         '''
         the original stream. (stored as a weakref but returned unwrapped)
 
-        >>> example = tree.makeExampleScore()
+        >>> example = tree.examples.makeExampleScore()
         >>> eTree = example.asTree()
         >>> eTree.source is example
         True
@@ -910,7 +910,7 @@ class OffsetTree(ElementTree):
         TRUE IF and ONLY if the
         .offset of the element matches the position in the tree -- thus it is very fast!
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True, groupOffsets=True)
 
         >>> score.flatten()[5] in scoreTree
@@ -941,7 +941,7 @@ class OffsetTree(ElementTree):
         r'''
         Gets elements by integer index or slice.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True, groupOffsets=True)
 
         >>> scoreTree[0]
@@ -1066,7 +1066,7 @@ class OffsetTree(ElementTree):
 
         Not an especially efficient way of using this beautiful tree object.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True, groupOffsets=True)
         >>> for x in scoreTree:
         ...     print(x)
@@ -1149,7 +1149,7 @@ class OffsetTree(ElementTree):
         with 3600 items took 500ms.  Creating the tree the first time
         was 40 seconds, so about an 80x speedup.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTimespans()
         >>> newTree = scoreTree.copy()
         >>> newTree
@@ -1441,7 +1441,7 @@ class OffsetTree(ElementTree):
         Creates a dictionary of offsets that have more than one element starting at that time,
         where the keys are offset times and the values are lists of elements at that moment.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = score.asTree(flatten=True, groupOffsets=True)
         >>> scoreTree
         <OffsetTree {20} (0.0 to 8.0) <music21.stream.Score exampleScore>>
