+++ music21/tree/verticality.py
@@ -581,7 +581,7 @@ class Verticality(prebase.ProtoM21Object):
         r'''
         Makes a Chord or Rest from this verticality and quarterLength.
 
-        >>> score = tree.makeExampleScore()
+        >>> score = tree.examples.makeExampleScore()
         >>> scoreTree = tree.fromStream.asTimespans(score, flatten=True,
         ...            classList=(note.Note, chord.Chord))
         >>> verticality = scoreTree.getVerticalityAt(4.0)
